request updates
you are able to send request
entites changed
reformed to one entity instead of 2

+++ ChatApp/backend/src/main/java/com/example/chatapp/Dto/UserDTO.java
@@ -7,7 +7,7 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class UserDTO {
 
-    private int userid;
+    private Long userid;
     private String username;
     private String email;
     private String password;
