+++ ChatApp/backend/src/main/java/com/example/chatapp/payload/response/UserResult.java
@@ -2,7 +2,7 @@ package com.example.chatapp.payload.response;
 
 public sealed interface UserResult permits UserResult.Success, UserResult.Failure {
 
-    public record Success(String username, int userId) implements UserResult {}
+    public record Success(String username, Long userId) implements UserResult {}
 
     public record Failure(String message) implements UserResult {}
 }
